Fixes slow startup: Closes bug #922237
Bug was due to the recent Events views, where photos are now sorted by
exposure date/time.  Old code was fetching and parsing the date/time out
of the Exiv2 date structure each time, which is expensive.  The date/time
is now cached inside the photo object for the duration of the app run.
When we have a persistent date store, we can store it there during the
import process.

This patch also fixes a couple of bugs found in the PhotoMetadata class.