Album Viewer now loads almost instantaneously: Closes bug #925672
Album Viewer was previously causing a lot of I/O loading images due to
the various components it held: the template pager, the grid view, and
a photo viewer instance, all of which pre-loaded various images as they
were initialized.  This patch reduces much of that I/O via caching and
late initialization.

The album preview is pokier than we will want right now.  This can be
fixed with larger thumbnails (bug #926328), as right now album previews
must use the master image for quality scaling.