Fixing signal emission/propagation from albums.  Fixes lp #931812.
It used to be that when photos were added to an album, the album would
erroneously fire several signals saying the page number changed.  Also, the
album collection model wasn't catching and dealing with the signal saying the
album page contents changed.  This fixes these issues and adds a signal on
albums for when the page contents are changed.  This all means QML now has what
it needs to automatically catch that the album thumbnails need to be redrawn.