Second commit for event timeline transition: Bug #915671
This improves the code, particularly by removing QmlEventMarker, a
stop-gap object that was made unnecessary by Event but not removed out
of lack of necessity.  More importantly, this commit now allows for the
transition to round-trip: the animation occurs both from overview to
timeline and timeline to overview by pressing an event card.