Animating the album opening/closing in the overview.  Fixes bug #915659.
Lots of tweaks to move everything into the right place to make this happen.
Also uncovered and fixed a bug or two in the AlbumPageViewer -- close() works
correctly now, and the clipping regions don't clip quite so aggressively.

This animation won't work incrementally because it's kind of faked using some
precise animation timing to get everything to coordinate so it looks good.
Doing it so it follows the mouse will be a whooooole 'nother piece of work.