Making dots in the page indicator "approximate."  Closes bug #937771.
Now each dot can have a number of pages associated with it, and clicking on it
turns to the first (though only if you aren't currently on one of its pages).
The pages are as evenly distributed as possible among the dots, but the last
dot always contains only the last page so you have quick access to the first
and last pages.  This distribution logic only kicks in when you hit the maximum
number of dots, 20.