Massive refactoring of album pages.
Used to be, a "page" was a full spread, like an open book.  Now a "page" is a
sheet of paper with a front and a back, that you can flip.  The AlbumPageViewer
has been refactored with this new definition.  This also makes the pages scale
up and down differently, so they look the same as the grow and shrink, which is
important with the new transitions coming up.

Couple problems: 1) the C++ code wasn't really changed enough, due to time
constraints.  We need a "page" to be half of a spread up and down the stack.
That's on the docket next.  As a consequence of that, 2) if you flip multiple
pages forward or backward in the viewer, you see the pages pop.  That'll be
fixed as soon along with #1.