Huge update to the page model for albums.
Now, one page is one side of a spread.  Albums always have an even number of
pages, so the covers always appear in the correct places.  There's a
distinction between "content pages", where photos can appear, and total pages,
which includes the covers.  There's one numbering scheme, so the cover is on
page 0, the first content page is page 1, etc.  There are helper properties to
help you deal with this.