Added an album open/close transition in the overview.  Closes bug #955297.
Updated AlbumPageContents to display the new album thumbnail frame image behind
it (and scale correctly) if desired.

Updated AlbumPreviewComponent to make use of the full AlbumPageViewer again
(and tweaked that so that it would work as a page vs. a spread), and put the
open/close transition back in the overview.  It's not perfect, but it gets the
job done for now.

Also, fixed a copy/paste error that had crept in with my previous refactorings.
The album page margins were all set to the top margin before; now they're using
the correct values.