Transition to close album when flipping to cover in viewer.  For bug #949407.
This is 90% of the functionality for 10% of the effort: it's not incremental.
As soon as you start to flip back toward the cover, the animation kicks off and
closes the album for you, ignoring further mouse movement.  This can be kind of
jarring if you don't expect it, just because of how small swipe threshold
distance is.

Also fixes a bug in SwipeArea where if you hid the area mid-swipe it wouldn't
reset its state properly when re-enabled.

Making this transition incremental will be a lot more work, so I'm leaving it
for a future commit.