Open/close album duration depends on where it's started.  Fixes bug #976715.
If you incrementally push the animation halfway, it now takes half as long to
finish.  If you finish it all the way, it finishes instantaneously.

Also fixed the text underneath previews in the album overview to display as
soon as you manually close them, if you push them all the way.