Tap a photo in Album View to view album in Photo Viewer: Closes bug #968387
Tapping a photo in the Album Viewer now launches the Photo Viewer, allowing
the entire album to be scrolled through.  When the user returns to the
Album Viewer, the album's current page will be updated to show the page
with the last seen photo.

Hopefully the hit-test code in AlbumSpreadViewer will be of some use with
bug #975347 (album viewer selection mode).