Cancel photo editing.  Fixes bug #1003707.
Now, the first time you make an edit on a photo, Gallery will back up the original in a
new .savepoints directory (whose name may change in the future).  When you cancel, it
simply reverts to the backed-up version (and when you hit done, it removes the backed-
up version just the same).