Fixing errors creating save points.  Fixes bug #1009700.
Now deletes the save point file before moving the file to the save point file,
so the move will succeed.

Also, renamed .savepoints to .temp to be clearer, and renamed thumbs to .thumbs
so we don't crowd up the Pictures directory with stuff only Gallery cares
about.  It's safe to delete your existing .savepoints and thumbs directories.