Revert to original.  Fixes bug #1009131.
Now there's a dummy "undo" menu in the photo editor with just the revert
option.  Every time you make an edit we now capture the original photo in a
folder named .original (in addition to the save point we were already
capturing).  When you revert to original, it moves the original back over the
"edited" photo and deletes any save points (so hitting cancel then will leave
you with the original).  This behavior can be tweaked.  Deleting the photo also
deletes the original.