Riffle effect when flipping pages.  Fixes bug #915672.
I rewrote the AlbumSpreadViewer to simplify it, and refactored out the new
AlbumPageFlipper component to handle the now more complex logic of animating
potentially multiple pages at once.  Now if you're in the album viewer on an
album with a number of pages, you can see several pages (5 at most) fly by if
you use the page indicator to skip some pages.  This can be tweaked as desired.