Close albums when paging past last page.  Fixes bug #949407.
Adding a new custom component AlbumCloser (largely copy/pasted from
AlbumOpener) to handle this custom animation.  When you try to page past the
last page in an album, at a point you trigger the (non-incremental) animation
that closes the album.  It's very similar to the animation when paging before
the first page, but the back page lifts up some before the animation kicks off
(which is probably desirable to duplicate in the front-cover case, but that's
for another ticket).

It ain't pretty (code-wise) but it gets the job done.  I'd like to find a way
eventually to re-combine the AlbumCloser and AlbumOpener, but that requires
more thought than I have time for now.