Change album page layout algorithm.  Fixes bug #968386 and bug #1023113.
We use an algorithm to pick a page layout that tries to preserve landscape or
portrait layout wherever possible.  In some preliminary testing, the "error"
rate (where we're cropping a portrait photo into a landscape frame or vice
versa) is much, much lower.

One outstanding issue: it can be slow.  We need to cache image dimensions a
little sooner.