Caching media size in DB.
This fixes the issue from the previous commit where it's slow to create albums
due to needing to look up photos' dimensions.  Now we cache photos' dimensions
as soon as they're loaded into the system (along with thumbnail generation).
There'll be a pause on the first run after this commit due to caching the
information, and then subsequent runs should be as fast as ever.