Tap tab bar to scroll to top.  Fixes bug #1025627.
Now in the overview, if you scroll down some, make the tab bar reappear, then
click the empty space in it (or the already selected tab), the view will scroll
up to the top.  This works in the event and album overview, and the event
timeline view.

Also found BinaryTabGroup and factored that out into Capetown.