Launch photo viewer from timeline.  Fixes bug #1026713.
Now when you click a photo in the event timeline, it launches the photo viewer
with only photos from that event.  Exiting the photo viewer uses the fade out
transition, which seems to be the most consistent way to handle some photos
being hidden in the timeline view, thus having no thumbnail to zoom out to.
Entering the photo editor will revert you back to the event overview when you
exit it.