Album pages no longer simply stretch when resizing.  Mostly fixes bug #1026704.
In the past, album pages were simply stretched up and down when changing sizes.
Now, we control the sizes of the margins independently and display the photos
at the propery aspect ratio inside the frames, which are changing size based
on margins and window size.

This also paves the way for the portrait mode album page view.

Note the new frame images are placeholders.  The ticket can be closed when we
get the correct assets.