Bottom photos in full view in selection mode.  Fixes bug #1036263.
In selection mode in the overview, the media selector, and also grid view in
the album viewer, you can now scroll down so that the bottom toolbar doesn't
obscure part of the last row of photos.  In the overview, when you exit
selection mode, the view is automatically scrolled back up a smidge so there's
no empty space at the bottom.