Massive restructuring of files.  Fixes bug #1021523.
New layout (please excuse my "ASCII" art):
  Capetown - generic QML utility components
  ├─Viewer (preexisting) - media viewer component to go into the SDK
  └─Widgets - UI that probably belongs in an SDK

  img - top-level or shared images (see below)

  js - new home for JavaScript files

  qml - Gallery's top-level views and transitions
  ├─Components - objects to display Gallery's core objects such as albums,
  │ │            photos, events, etc.
  │ └─AlbumInternals - internal stuff necessary to make albums behave like
  │                    albums
  ├─Utility - misc. reusable items specific to Gallery
  └─Widgets - UI that will be styled from the SDK

Also, inside each directory containing QML, there may be an img directory.
This contains images specific to that particular layer of QML code.  This isn't
ideal, since images may be used by multiple layers (plus it can be confusing if
an image's path is defined in one layer, but the component is loaded in a
separate layer, as is the case with menu icons).  So, many images had to end up
in the top-level, shared img directory out of necessity.  The alternative was
to have long strings of "../../.." in image paths, which seemed even worse.
Either way, a better solution is needed here, but that'll be a future project.