Support more image formats.  Fixes bug #1043974.
We now support anything Qt can read and write.  We store the rotation in the
database (in addition to the file metadata where possible).  Metadata is only
written to files of certain whitelisted formats (and within that, there's an
additional whitelist of files we can store the rotation in, due to different
types of metadata being valid in different types of files).