Tweak chrome fade durations.  Fixes bug #1044567.
I've tweaked many UI animations, mostly in the photo viewer, but also for the
event timeline transition.  Right now I'm not very fond of the new timeline
transition, but it may grow on me.

I've also tried to lay the groundwork for when the SDK provides us some values
to use for UI animation durations.