Demo incremental timeline transition.  Big step toward bug #943731.
This makes the event timeline transition incremental.  For now, we just display
a green bar in the toolbar that lets you swipe to control the transition, and
there are some kinks with how that works.  In the future of course it will be
multitouch, and the swipe kinks should go away.

If the green bar is undesirable, it can be ripped out easily.  It's just there
to demonstrate the capabilities.